// generated/src/app/api-models/line-of-business-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class LineOfBusinessMaster {
  lineOfBusiness: string;
  description: string;
  name: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  phoneNumber: string;
  faxNumber: string;
  pricePriorityRule: string;
  providerDefault: number;
  pcpRequired: string;
  eobPrintFlag: string;
  subDepEditSwitch: string;
  claimDupRule: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  performAuthClaimMatch: string;
  authClaimMatchDaysBefore: number;
  authClaimMatchDaysAfter: number;
  coveringProvInd: string;
  country: string;
  inclmSwitch: string;
  performAuthWaive: string;
  authLevel: string;
  allowedGreaterThanBilled: string;
  targetRevCodeEditFlg: string;
  targetRevAction: string;
  targetRevReason: string;
  parReasonCode: string;
  nonParReasonCode: string;
  indPcpMaxEnrollLmt: number;
  indPcpThreshold: number;
  seqDefPcp: number;
  seqFailPcp: number;
  idcardReinstDays: number;
  familyAffiliation: string;
  useApcEditor: string;
  editType: string;
  useApcGrouper: string;
  grouperAction: string;
  grouperReasonCode: string;
  useApcPricer: string;
  pricerAction: string;
  pricerReasonCode: string;
  pricerNotCovReason: string;
  interestDeterminant: string;
  useResetDateFlag: string;
  intDscntPaySub: string;
  calcIntDscnt: string;
  applyUncleanInd: string;
  userDefined1: string;
  userDefined2: string;
  userDate1: Date;
  userDate2: Date;
  waiveMatchOrder: string;
  certifiedFlg: string;
  ageFrom: number;
  ageThrough: number;
  siteFlg: string;
  mcIndicator: string;
  dateOfDeathRule: string;
  penApplyUncleanInd: string;
}